pyproject.toml
[project]
name = "sign-language-translator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "ctranslate2>=4.0.0",
    "edge-tts>=7.2.3",
    "faster-whisper>=1.2.1",
    "google-genai>=1.52.0",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "librosa>=0.11.0",
    "mediapipe>=0.10.14",
    "numpy>=2.3.5",
    "protobuf<=4.25.3",
    "pyaudio>=0.2.14",
    "pygame>=2.6.1",
    "python-dotenv>=1.2.1",
    "setuptools<80",
    "speechrecognition>=3.14.4",
    "spotipy>=2.25.2",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchvision>=0.24.1",
]
